+++ PureInvoke/Functions/Invoke-KernelGetVolumePathName.ps1
@@ -1,8 +1,22 @@
 
 function Invoke-KernelGetVolumePathName
 {
+    <#
+    .SYNOPSIS
+    Calls the kernel32.dll libary's `GetVolumePathName` function.
+
+    .DESCRIPTION
+    The `Invoke-KernelGetVolumePathName` function calls the kernel32.dll libary's `GetVolumePathName` function which
+    gets the volume mount point of the path. Pass the path to the `Path` parameter.
+
+    .EXAMPLE
+    Invoke-KernelGetVolumePathName -Path $path
+
+    Demonstrates how to call this function.
+    #>
     [CmdletBinding()]
     param(
+        # The path whose volume mount point to get.
         [Parameter(Mandatory)]
         [String] $Path
     )
@@ -10,20 +24,17 @@ function Invoke-KernelGetVolumePathName
     Set-StrictMode -Version 'Latest'
     Use-CallerPreference -Cmdlet $PSCmdlet -Session $ExecutionContext.SessionState
 
-    [Marshal]::SetLastPInvokeError([PureInvoke.ErrorCode]::Ok)
-    [Marshal]::SetLastSystemError([PureInvoke.ErrorCode]::Ok)
     # Need to call GetFullPathName to get the size of the necessary buffer.
     $sbPath = [Text.StringBuilder]::New([PureInvoke.Kernel32]::MAX_PATH)
     $cchPath = [UInt32]$sbPath.Capacity; # in/out character-count variable for the WinAPI calls.
-    # Get the volume (drive) part of the target file's full path (e.g., @"C:\")
     $result = [PureInvoke.Kernel32]::GetVolumePathName($Path, $sbPath, $cchPath)
-    [PureInvoke.ErrorCode] $errCode = [Marshal]::GetLastWin32Error()
-    $msg = "[Kernel32]::GetVolumePathName(""${Path}"", ""${sbPath}"", ${cchPath})  return ${result}  " +
+    # Get the volume (drive) part of the target file's full path (e.g., @"C:\")
+    $errCode = [Marshal]::GetLastWin32Error()
+    $msg = "[Kernel32]::GetVolumePathName(""${Path}"", [out] ""${sbPath}"", ${cchPath})  return ${result}  " +
            "GetLastError() ${errCode}"
     Write-Debug $msg
-    if (-not $result)
+    if (-not $result -and -not (Assert-Win32Error -ErrorCode $errCode))
     {
-        Write-Win32Error -ErrorCode $errCode
         return
     }